requirements.txt
opencv-python
opencv-contrib-python
numpy
streamlit
matplotlib
plotly
pillow
